Modify Rust client to ignore unknown plugins
* Add full collection deserialization.

* Remove fetch_plugins function.
