+++ clients/rust/src/hooked/collection.rs
@@ -1,13 +1,13 @@
 use borsh::BorshSerialize;
 
 use crate::{
-    accounts::{BaseAssetV1, PluginHeaderV1},
-    registry_records_to_plugin_list, Asset, PluginRegistryV1Safe,
+    accounts::{BaseCollectionV1, PluginHeaderV1},
+    registry_records_to_plugin_list, Collection, PluginRegistryV1Safe,
 };
 
-impl Asset {
-    pub fn deserialize_asset(data: &[u8]) -> Result<Self, std::io::Error> {
-        let base = BaseAssetV1::from_bytes(data)?;
+impl Collection {
+    pub fn deserialize_collection(data: &[u8]) -> Result<Self, std::io::Error> {
+        let base = BaseCollectionV1::from_bytes(data)?;
         let base_data = base.try_to_vec()?;
         let (plugin_header, plugin_list) = if base_data.len() != data.len() {
             let plugin_header = PluginHeaderV1::from_bytes(&data[base_data.len()..])?;
@@ -31,17 +31,17 @@ impl Asset {
 
     #[inline(always)]
     pub fn from_bytes(data: &[u8]) -> Result<Self, std::io::Error> {
-        Self::deserialize_asset(data)
+        Self::deserialize_collection(data)
     }
 }
 
-impl<'a> TryFrom<&solana_program::account_info::AccountInfo<'a>> for Asset {
+impl<'a> TryFrom<&solana_program::account_info::AccountInfo<'a>> for Collection {
     type Error = std::io::Error;
 
     fn try_from(
         account_info: &solana_program::account_info::AccountInfo<'a>,
     ) -> Result<Self, Self::Error> {
         let data: &[u8] = &(*account_info.data).borrow();
-        Self::deserialize_asset(data)
+        Self::deserialize_collection(data)
     }
 }
