+++ clients/rust/src/hooked/mod.rs
@@ -7,8 +7,10 @@ pub use advanced_types::*;
 pub mod asset;
 pub use asset::*;
 
-use borsh::{BorshDeserialize, BorshSerialize};
+pub mod collection;
+pub use collection::*;
 
+use borsh::{BorshDeserialize, BorshSerialize};
 use std::{cmp::Ordering, mem::size_of};
 
 use crate::{
@@ -17,6 +19,24 @@ use crate::{
     types::{Key, Plugin, PluginType, RegistryRecord},
 };
 use solana_program::account_info::AccountInfo;
+
+impl PluginType {
+    // Needed to determine if a plugin is a known or unknown type.
+    pub fn from_u8(n: u8) -> Option<PluginType> {
+        match n {
+            0 => Some(PluginType::Royalties),
+            1 => Some(PluginType::FreezeDelegate),
+            2 => Some(PluginType::BurnDelegate),
+            3 => Some(PluginType::TransferDelegate),
+            4 => Some(PluginType::UpdateDelegate),
+            5 => Some(PluginType::PermanentFreezeDelegate),
+            6 => Some(PluginType::Attributes),
+            7 => Some(PluginType::PermanentTransferDelegate),
+            8 => Some(PluginType::PermanentBurnDelegate),
+            _ => None,
+        }
+    }
+}
 
 impl From<&Plugin> for PluginType {
     fn from(plugin: &Plugin) -> Self {
